Moved initial testing routes to users.py and created User class model
